LayoutTests/http/tests/links/redirect-301-no-inital-fragments-expected.txt


PASS: document.getElementById("first_load").contentWindow.location.hash should be and is. PASS: document.getElementById("second_load").contentWindow.location.hash should be #key=bar and is.
